requirements.txt
ollama
Pillow
PyYAML
